move service resource logic into _templates.services
