+++ cli/src/accretion_cli/_templates/services/__init__.py
@@ -1,0 +1,1 @@
+""""""
