+++ cli/src/accretion_cli/_templates/services/stepfunctions.py
@@ -4,7 +4,7 @@ from typing import Dict, Iterable
 
 from troposphere import AWSObject, Sub, Tags, Template, awslambda, sns, stepfunctions
 
-from ._iam import invoke_statement_from_lambdas, publish_statement_from_topics, step_functions_role
+from .iam import invoke_statement_from_lambdas, publish_statement_from_topics, step_functions_role
 
 
 def _resource_to_sub_reference(resource: AWSObject) -> str:
