+++ cli/test/vectors/replication_listener_template.json
@@ -21,7 +21,7 @@
                 "Tags": [
                     {
                         "Key": "Accretion",
-                        "Value": "ArtifactBuilder"
+                        "Value": "ReplicationListener"
                     }
                 ]
             }
@@ -98,7 +98,7 @@
                 "Tags": [
                     {
                         "Key": "Accretion",
-                        "Value": "ArtifactBuilder"
+                        "Value": "ReplicationListener"
                     }
                 ]
             }
@@ -179,7 +179,7 @@
                 "Tags": [
                     {
                         "Key": "Accretion",
-                        "Value": "ArtifactBuilder"
+                        "Value": "ReplicationListener"
                     }
                 ]
             }
@@ -274,7 +274,7 @@
                 "Tags": [
                     {
                         "Key": "Accretion",
-                        "Value": "ArtifactBuilder"
+                        "Value": "ReplicationListener"
                     }
                 ]
             }
@@ -389,7 +389,7 @@
                 "Tags": [
                     {
                         "Key": "Accretion",
-                        "Value": "ArtifactBuilder"
+                        "Value": "ReplicationListener"
                     }
                 ]
             }
@@ -538,7 +538,7 @@
                 "Tags": [
                     {
                         "Key": "Accretion",
-                        "Value": "ArtifactBuilder"
+                        "Value": "ReplicationListener"
                     }
                 ]
             }
